auth

+++ app/api/api_v1/endpoints/message_api.py
@@ -7,6 +7,7 @@ from app.services.chat_service import ChatService
 from typing import Annotated
 import logging
 from app.models.message_model import *
+from app.services.auth_service import auth_wrapper
 
 logger = logging.getLogger()
 router = APIRouter()
@@ -16,7 +17,12 @@ router = APIRouter()
 async def get_history(
                     chatid: Annotated[int | None, Query()]=None,
                     chat_service: ChatService = Depends(), 
-                    session:AsyncSession=Depends(get_session)):
+                    session:AsyncSession=Depends(get_session),
+                    user=Depends(auth_wrapper)
+                    ):
+    if user == 'fail':
+        raise HTTPException(status_code=401, detail="Invalid token")
+
     print("GET HISTORY OF ", chatid)
     history = await chat_service.get_chat_history(chatid, session)
     return history
@@ -26,7 +32,12 @@ async def get_history(
 async def send_message(
                     message: MessageCreate,
                     chat_service: ChatService = Depends(),
-                    session: AsyncSession = Depends(get_session)):
+                    session: AsyncSession = Depends(get_session),
+                    user=Depends(auth_wrapper)
+                    ):
+    if user == 'fail':
+        raise HTTPException(status_code=401, detail="Invalid token")
+
     print("SEND MESSAGE TO ", message)
     result = await chat_service.send_message(message, session)
     print('Answer: ', result)
@@ -36,7 +47,12 @@ async def send_message(
 async def delete_message(
                     chatid: int = Path(..., title="The ID of the chat"),
                     chat_service: ChatService = Depends(),
-                    session: AsyncSession = Depends(get_session)):
+                    session: AsyncSession = Depends(get_session),
+                    user=Depends(auth_wrapper)
+                    ):
+    if user == 'fail':
+        raise HTTPException(status_code=401, detail="Invalid token")
+
     print("DELETE MESSAGE FROM ", chatid)
     result = await chat_service.delete_message(chatid, session)
     print('Answer: ', result)
